requirements.txt
django-prometheus==2.2.0
django==4.1.3
psycopg2==2.9.5
ipython==8.6.0
requests==2.28.1
djangorestframework==3.14.0
pytz==2021.3
django-cors-headers==3.10.0
django-filter==22.1
